statesync: Move into internal
